refactor(frontend): extract auto-spider effect into useAutoSpider hook (#293)
MapCanvas was ~1100 lines; the auto-spider reconciler (post-Spider v2)
was ~200 lines of dense projection / grouping / fan-position math +
source/layer lifecycle. Extract to use-auto-spider.ts so MapCanvas
just consumes the returned stacks and renders the Markers.

Behavior-preserving — all existing auto-spider tests (5-co-located,
no-overlap, silhouettes-empty short-circuit, listener cleanup,
second-idle no-oscillation, tile-boundary dedupe) pass unchanged.
